+++ tests/lib/rules/no-unused-state.js
@@ -989,16 +989,16 @@ eslintTester.run('no-unused-state', rule, {
     semver.satisfies(tsEslintVersion, '>= 5') ? {
       code: `
         interface Props {}
-        
+
         interface State {
           flag: boolean;
         }
-        
+
         export default class RuleTest extends React.Component<Props, State> {
           readonly state: State = {
             flag: false,
           };
-        
+
           static getDerivedStateFromProps = (props: Props, state: State) => {
             const newState: Partial<State> = {};
             if (!state.flag) {
@@ -1030,7 +1030,7 @@ eslintTester.run('no-unused-state', rule, {
         class KarmaRefundPillComponent extends GenericPillComponent {
           renderContent = () => {
             const { action } = this.props
-        
+
             return (
               <Box fontSize={[1]} mx={[2]} minWidth="10px" minHeight="26px" alignItems="center">
                 <FormattedText
@@ -1059,6 +1059,40 @@ eslintTester.run('no-unused-state', rule, {
           static getDerivedStateFromProps = ComposeComponent.getDerivedStateFromProps;
           render() { return <div />; }
         }
+      `,
+      features: ['class fields'],
+      parserOptions: {
+        sourceType: 'module',
+      },
+    },
+    {
+      code: `
+        import React, { PureComponent } from 'react';
+
+        class TestNoUnusedState extends React.Component {
+          constructor(props) {
+            super(props);
+            this.state = {
+              id: null,
+            };
+          }
+
+          static getDerivedStateFromProps = (props, state) => {
+            if (state.id !== props.id) {
+              return {
+                id: props.id,
+              };
+            }
+
+            return null;
+          };
+
+          render() {
+            return <h1>{this.state.id}</h1>;
+          }
+        }
+
+        export default TestNoUnusedState;
       `,
       features: ['class fields'],
       parserOptions: {